+++ apps/wallet/src/ui/app/hooks/deepbook/index.ts
@@ -5,3 +5,4 @@ export * from './useDeepBookConfigs';
 export * from './useRecognizedCoins';
 export * from './useDeepbookPools';
 export * from './useBalanceConversion';
+export * from './useGetEstimate';
