+++ apps/wallet/src/ui/app/hooks/index.ts
@@ -17,7 +17,6 @@ export { useQueryTransactionsByAddress } from './useQueryTransactionsByAddress';
 export { useGetTransferAmount } from './useGetTransferAmount';
 export { useOwnedNFT } from './useOwnedNFT';
 export { useSortedCoinsByCategories } from './useSortedCoinsByCategories';
-export * from './useDeepBook';
 export * from './useTransactionData';
 export * from './useActiveAddress';
 export * from './useGetAllCoins';
